Added error message when there are no file fields in datasource that need them. Fixed a bug where redbox uri's would include the search string. Added fix to ca_model where bool values of 'false' were being interpreted by collander boolean as True - when rendering without validation. Minor text updates in project.py. Untested implementation of cleanup_uploads.py hid the example field from custom fields of type file. Added a message to data manager with a link to the data portal. Allowed users to delete projects before they are submitted, and admins can delete before approved.

+++ jcudc24provisioning/scripts/cleanup_uploads.py
@@ -26,6 +26,7 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 import sys
 from jcudc24provisioning.models import DBSession
+from jcudc24provisioning.models.project import Attachment, MethodAttachment
 import os
 from pyramid.paster import get_appsettings
 from sqlalchemy import engine_from_config
@@ -62,7 +63,7 @@ def cleanup_uploads(config_uri):
         sys.exit(1)
 
     for f_name in os.listdir(uploads_folder):
-        session.query()
-
-        os.remove(f_name)
+        if session.query(Attachment).filter_by(attachment=f_name).count() + \
+           session.query(MethodAttachment).filter_by(attachment=f_name).count() == 0:
+            os.remove(f_name)
